# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Fission HTTP Server API
  description: |
    REST API for managing Fission resources on Kubernetes.
    This API wraps all fission-cli commands and provides HTTP endpoints for:
    - Functions (create, list, get, update, delete, test, logs, pods, metadata)
    - Environments (create, list, get, update, delete, pods)
    - Packages (create, list, get source/deploy/info, update, delete, rebuild)
    - HTTP Triggers (create, list, get, update, delete)
    - Time Triggers (create, list, delete)
    - MQ Triggers (create, list, delete)
    - KubeWatches (create, list, delete)
    - Version information
    - System check
    - Plugins
  version: 2.0.0
  contact:
    name: Fission HTTP Server
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.example.com
    description: Production server (example)
paths:
...
  /api/v1/httptriggers:
    get:
      summary: List all HTTP triggers
      description: Lists all HTTP triggers.
      operationId: listHTTPTriggers
      tags:
        - HTTP Triggers
      parameters:
        - name: namespace
          in: query
          required: false
          schema:
            type: string
        - name: all_namespaces
          in: query
          required: false
          schema:
            type: boolean
        - name: function
          in: query
          required: false
          schema:
            type: string
            description: Filter by function name
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  output:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create an HTTP trigger
      description: Creates a new HTTP trigger.
      operationId: createHTTPTrigger
      tags:
        - HTTP Triggers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                function:
                  type: string
                url:
                  type: string
                method:
                  type: string
                prefix:
                  type: string
      responses:
        '201':
          description: HTTP trigger created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  output:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/httptriggers/{name}:
    get:
      summary: Get HTTP trigger details
      description: Retrieves details about a specific HTTP trigger.
      operationId: getHTTPTrigger
      tags:
        - HTTP Triggers
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  output:
                    type: string
        '404':
          description: HTTP trigger not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update an HTTP trigger
      description: Updates an existing HTTP trigger.
      operationId: updateHTTPTrigger
      tags:
        - HTTP Triggers
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                function:
                  type: string
                url:
                  type: string
                method:
                  type: string
      responses:
        '200':
          description: HTTP trigger updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  name:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete an HTTP trigger
      description: Deletes an HTTP trigger.
      operationId: deleteHTTPTrigger
      tags:
        - HTTP Triggers
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: namespace
          in: query
          required: false
          schema:
            type: string
        - name: ignore_not_found
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: HTTP trigger deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  name:
                    type: string
        '404':
          description: HTTP trigger not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Time Trigger endpoints
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      description: Error response
      properties:
        error:
          type: string
          description: Error message
          example: "An error occurred"
      required:
        - error